frontend/core/pipelines/canny-edge-demo/canny-edge-k2o7j05im846/requirements.txt
Pillow==9.5.0
opencv-python-headless==4.5.5.62
numpy==1.26.4
